put all enums together in a file
