+++ lib/models/all_enums.dart
@@ -1,0 +1,44 @@
+/// Enums for consistent data
+enum SetStatus { draft, complete }
+
+enum BoardStatus { draft, complete }
+
+enum QuestionStatus { draft, complete }
+
+enum RoomStatus { waiting, active, questionActive, ended }
+
+enum DifficultyLevel { easy, medium, hard }
+
+enum PredefinedTags {
+  architecture,
+  arts,
+  astronomy,
+  biology,
+  business,
+  civics,
+  words,
+  entertainment,
+  fashion,
+  foodAndDrinks,
+  general,
+  geography,
+  history,
+  india,
+  literature,
+  logos,
+  maths,
+  movies,
+  mythology,
+  other,
+  personal,
+  politics,
+  popCulture,
+  science,
+  songs,
+  sports,
+  technology,
+  us,
+  videoGames,
+  wordplay,
+  world,
+}
